linux/requirements.txt
qwen-asr>=0.0.6
torch>=2.0
sounddevice>=0.4
numpy>=1.24
pynput>=1.7
pystray>=0.19
Pillow>=9.0
httpx>=0.24
PyYAML>=6.0
